# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Yend2 Awesome App",
        "description": "API for Yend2 Awesome App. Environment: development.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/push/sources": {
            "post": {
                "tags": [
                    "Push Notifications"
                ],
                "summary": " (管理员) 创建新的推送来源",
                "description": "创建一个新的推送来源。例如，配置一个钉钉机器人或一个通用的 Webhook 接收点。\n仅限超级管理员操作。",
                "operationId": "create_push_source_api_v1_push_sources_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PushSourceCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PushSourcePublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Push Notifications"
                ],
                "summary": " (管理员或认证用户) 获取推送来源列表",
                "description": "获取推送来源列表。\n- 普通认证用户或匿名用户通常只能看到激活的来源。\n- 管理员可以通过 `only_active=False` 查看所有来源。",
                "operationId": "read_push_sources_api_v1_push_sources_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 200,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "only_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "仅返回激活的推送来源 (对普通用户默认为True)",
                            "default": true,
                            "title": "Only Active"
                        },
                        "description": "仅返回激活的推送来源 (对普通用户默认为True)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PushSourcePublic"
                                    },
                                    "title": "Response Read Push Sources Api V1 Push Sources Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/push/sources/{source_id}": {
            "get": {
                "tags": [
                    "Push Notifications"
                ],
                "summary": " (管理员) 获取指定推送来源详情",
                "description": "获取特定推送来源的详细信息。\n仅限超级管理员操作（因为可能包含敏感配置的引用）。",
                "operationId": "read_push_source_details_api_v1_push_sources__source_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "source_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Source Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PushSourcePublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Push Notifications"
                ],
                "summary": " (管理员) 更新指定推送来源",
                "description": "更新特定推送来源的信息。\n仅限超级管理员操作。",
                "operationId": "update_push_source_api_v1_push_sources__source_id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "source_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Source Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PushSourceUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PushSourcePublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Push Notifications"
                ],
                "summary": " (管理员) 删除指定推送来源",
                "description": "删除特定的推送来源。\n相关的订阅和（可能）消息也会被级联删除（取决于模型定义）。\n仅限超级管理员操作。",
                "operationId": "delete_push_source_api_v1_push_sources__source_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "source_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Source Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PushSourceCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Name",
                        "description": "推送来源的易记名称"
                    },
                    "source_type": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Source Type",
                        "description": "来源类型 (如 'dingtalk', 'webhook', 'email')"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "推送来源的描述 (可选)"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "description": "此推送来源是否启用，默认为 True",
                        "default": true
                    },
                    "config": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Config",
                        "description": "来源的特定配置 (JSON对象)，例如 webhook URL、API 密钥等"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "name",
                    "source_type"
                ],
                "title": "PushSourceCreate",
                "description": "创建新推送来源时使用的 Schema。"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}